+++ packages/ratel_generator/test/support/stub_packages.dart
@@ -1,5 +1,6 @@
 const stubPackages = <String, String>{
   'ratel|lib/annotations/annotations.dart': _annotations,
+  'ratel|lib/core/request_context.dart': _requestContext,
   'ratel|lib/http/handler.dart': _handler,
   'ratel_orm|lib/annotations.dart': _ormAnnotations,
   'ratel_orm|lib/ratel_orm.dart': _orm,
@@ -99,6 +100,10 @@ class CookieParam {
   final String name;
   const CookieParam(this.name);
 }
+''';
+
+const _requestContext = '''
+class RequestContext {}
 ''';
 
 const _handler = '''